requirements.txt
yara-python == 4.1.3

